requirements.txt
discord.py
env-file
pymongo~=3.11.3